requirements.txt
bcrypt==2.0.0
beautifulsoup4==4.4.1
botocore==1.4.8
cffi==1.5.2
cloudinary==1.3.2
docutils==0.12
google-api-python-client==1.5.3
hiredis==0.2.0
jmespath==0.9.0
mock==1.3.0
pbr==1.8.1
pycparser==2.14
python-dateutil==2.4.2
PyYAML==3.12
redis==2.10.5
six==1.10.0
sockjs-tornado==1.0.2
tornado==4.3
tornado-redis==2.4.18
tornado-smtpclient==0.1.5
wheel==0.29.0